[feat] 리뷰 내역 반영
